Refactor routing logic by moving it to a new router.ts module, simplifying index.ts and improving code organization; update server initialization to use the new router.
